+++ qcelemental/models/v2/common_models.py
@@ -170,6 +170,4 @@ def check_convertible_version(ver: int, error: str):
         raise ValueError(f"QCSchema {error} version={version} does not exist for conversion.")
 
 
-qcschema_torsion_drive_input_default = "qcschema_torsion_drive_input"
-qcschema_torsion_drive_output_default = "qcschema_torsion_drive_output"
 qcschema_molecule_default = "qcschema_molecule"